User story 5 auth

+++ app/controllers/wikis_controller.rb
@@ -1,6 +1,6 @@
 class WikisController < ApplicationController
   def index
-    @wikis = Wiki.all
+    @wikis = policy_scope(Wiki).all
   end
 
   def show
@@ -9,11 +9,13 @@ class WikisController < ApplicationController
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @wiki = Wiki.new(wiki_params)
-
+    @wiki.user = current_user
+    authorize @wiki
     if @wiki.save
       redirect_to wiki_path(@wiki)
       flash[:notice] = "Your Wiki has been created"
@@ -24,11 +26,12 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
-
+    authorize @wiki
     if @wiki.update_attributes(wiki_params)
       flash[:notice] = "Wiki updated"
       redirect_to @wiki
@@ -39,7 +42,8 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
-
+    #@wiki = policy_scope(Wiki).find(params[:id])
+    authorize @wiki
     if @wiki.destroy
       flash[:notice] = "Wiki deleted successfully"
       redirect_to @wiki
